requirements.txt
mcp[cli]>=1.5.0
httpx>=0.24.0
beautifulsoup4>=4.10.0
caldav>=0.8.0
python-dotenv>=0.19.0
